weekOneJavaAssignment/src/main/java/SchoolSystem/Classes.java
package SchoolSystem;

import java.util.ArrayList;
import java.util.List;
import Employees.Teacher;

public class Classes {
    private String className;
    private List<Student> students;
    private Teacher teacher;

    public Classes(String className, Teacher teacher) {
        this.className = className;
        this.teacher = teacher;
        this.students = new ArrayList<>();
    }

    public String getClassName() {
        return className;
    }

    public void setClassName(String className) {
        this.className = className;
    }

    public List<Student> getStudents() {
        return students;
    }

    public void setStudents(List<Student> students) {
        this.students = students;
    }

    public Teacher getTeacher() {
        return teacher;
    }

    public void setTeacher(Teacher teacher) {
        this.teacher = teacher;
    }

    public void addStudent(Student student) {
        students.add(student);
    }

}
